Move native methods to a separate directory
