+++ src/reanimated2/index.ts
@@ -91,10 +91,10 @@ export type {
   EasingFactoryFn,
 } from './Easing';
 export { Easing } from './Easing';
-export { measure } from './measure';
-export { dispatchCommand } from './dispatchCommand';
-export { scrollTo } from './scrollTo';
-export { setGestureState } from './setGestureState';
+export { measure } from './nativeMethods/measure';
+export { dispatchCommand } from './nativeMethods/dispatchCommand';
+export { scrollTo } from './nativeMethods/scrollTo';
+export { setGestureState } from './nativeMethods/setGestureState';
 export { setNativeProps } from './SetNativeProps';
 export type { ParsedColorArray } from './Colors';
 export { isColor, processColor, convertToRGBA } from './Colors';
